Rename controller, clear input
