+++ app/views/conversations/_user_bot_msg.html.erb
@@ -1,12 +1,14 @@
-<% @conversation.user_messages.each_with_index do |user_msg, index| %>
-    <div class="user-msg">
-      <%= user_msg.content %>
-    </div>
-    <% next_bot_message = @conversation.bot_messages.find { |bot_msg| bot_msg.created_at > user_msg.created_at } %>
-    <% if next_bot_message %>
-      <div class="bot-msg">
-        <%= next_bot_message.content %>
+<div class="msgs" data-insert-msg-target="conversation">
+  <% @conversation.user_messages.each_with_index do |user_msg, index| %>
+      <div class="user-msg">
+        <%= user_msg.content %>
       </div>
+      <% next_bot_message = @conversation.bot_messages.find { |bot_msg| bot_msg.created_at > user_msg.created_at } %>
+      <% if next_bot_message %>
+        <div class="bot-msg">
+          <%= next_bot_message.content %>
+        </div>
+      <% end %>
+        <%= render "conversations/bot_audio_tag", bot_message: next_bot_message  %>
     <% end %>
-      <%= render "bot_audio_tag", bot_message: next_bot_message  %>
-  <% end %>
+</div>
