+++ app/views/conversations/show.html.erb
@@ -1,31 +1,35 @@
-<div class="chat-window"  id="chat">
-  <div class="chat-topic">
-    <h3><%= @conversation.topic.name %></h3>
-    <p><%= @conversation.topic.level %> level</p>
-    <button data-action="click->conversation-window-scrolling#scroll"></button>
-  </div>
-  <div class="bot-msg">
-    <%= @conversation.bot_messages.order('created_at ASC').first.content %>
-    <%= render "bot_audio_tag", bot_message: @conversation.bot_messages.first %>
-  </div>
+<div data-controller="insert-msg">
+  <div class="chat-window"  id="chat">
+    <div class="chat-topic">
+      <h3><%= @conversation.topic.name %></h3>
+      <p><%= @conversation.topic.level %> level</p>
+    </div>
+    <div class="bot-msg">
+      <%= @conversation.bot_messages.order('created_at ASC').first.content %>
+      <%= render "bot_audio_tag", bot_message: @conversation.bot_messages.first %>
+    </div>
 
-   <%= render "user_bot_msg", user_messages: @conversation.user_messages %>
-   
-</div>
-<div class = "input-with-btns" data-controller="audio-recording conversation-window-scrolling">
-  <div class = "input-field">
-
-    <%= simple_form_for [@conversation, @user_message], html: {class: "d-flex"} do |f| %>
-        <%= f.input :content, label: false, placeholder: "Press and hold the record button to start recording a message", class: "user-msg-field", wrapper_html: {class: "flex-grow-1"}, input_html: { data: { target: 'audio-recording.input'}} %>
-        <%= f.button :button, id: "send-msg", data: { action: "turbo:submit-end->conversation-window-scrolling#scroll"} do %>
-          <i class="fa-solid fa-paper-plane"></i>
-        <% end %>
-    <% end %>
-
+    <%= render "user_bot_msg", user_messages: @conversation.user_messages %>
 
   </div>
-  <div class="record" >
-    <button class="record-btn" data-audio-recording-target="recordButton" data-action="mousedown->audio-recording#startRecording mouseup->audio-recording#stopRecording"><i class="fa-solid fa-microphone"></i></button>
+  <div class = "input-with-btns" data-controller="audio-recording">
+    <div class = "input-field">
+
+      <%= simple_form_for [@conversation, @user_message], html: {class: "d-flex", data: {
+        insert_msg_target: "form",
+        action: "submit->insert-msg#scroll"
+    }} do |f| %>
+          <%= f.input :content, label: false, placeholder: "Press and hold the record button to start recording a message", class: "user-msg-field", wrapper_html: {class: "flex-grow-1"}, input_html: { data: { target: 'audio-recording.input insert-msg.input'}} %>
+          <%= f.button :button, id: "send-msg" do %>
+            <i class="fa-solid fa-paper-plane"></i>
+          <% end %>
+      <% end %>
+
+
+    </div>
+    <div class="record" >
+      <button class="record-btn" data-audio-recording-target="recordButton" data-action="mousedown->audio-recording#startRecording mouseup->audio-recording#stopRecording"><i class="fa-solid fa-microphone"></i></button>
+    </div>
+    <%=link_to "End conversation", end_conversation_path(@conversation), data: { 'turbo-method': :patch }, class:"end-btn"%>
   </div>
-  <%=link_to "End conversation", end_conversation_path(@conversation), data: { 'turbo-method': :patch }, class:"end-btn"%>
 </div>
